+++ Coditech.Project/Coditech.API.Model.Custom/DBTM/DBTMCamp/DBTMCampActivityModel.cs
@@ -3,7 +3,7 @@
     public class DBTMCampActivityModel : BaseModel
     {
         public long DBTMCampActivityId { get; set; }
-        public long DBTMCampMasterId { get; set; }
+        public int DBTMCampMasterId { get; set; }
         public int DBTMTestMasterId { get; set; }
         public string TestName { get; set; }
         public bool IsAssociated { get; set; }
